% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuillInput.R
\name{QuillInput}
\alias{QuillInput}
\title{Create a Quill Input Widget}
\usage{
QuillInput(inputId, default = "")
}
\arguments{
\item{inputId}{The input ID.}
}
\value{
A Quill input widget.
}
\description{
This function creates a Quill input widget for use in Shiny applications.
}
