% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuillInput.R
\name{updateQuillInput}
\alias{updateQuillInput}
\title{Update a Quill Input Widget}
\usage{
updateQuillInput(session, inputId, value, configuration = NULL)
}
\arguments{
\item{session}{The Shiny session.}

\item{inputId}{The input ID.}

\item{value}{The initial value.}

\item{options}{A list of options passed to Quill.}
}
\value{
A Quill input widget.
}
\description{
This function updates a Quill input widget for use in Shiny applications.
}
