% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuillInput.R
\name{QuillInput}
\alias{QuillInput}
\title{Create a Quill Input Widget}
\usage{
QuillInput(inputId, value = "", configuration = NULL, options = NULL)
}
\arguments{
\item{inputId}{The input ID.}

\item{value}{The initial value (default is an empty string).}

\item{configuration}{A list of configuration options passed to the JavaScript component.}

\item{options}{A list of options (can be empty if not used).}
}
\value{
A Shiny input widget.

A Shiny input widget.
}
\description{
Create a Quill Input Widget
}
